Made a working one level game
