+++ README.md
@@ -1,8 +1,6 @@
 # LearningType
 
-Welcome to your new gem! In this directory, you'll find the files you need to be able to package up your Ruby library into a gem. Put your Ruby code in the file `lib/learning_type`. To experiment with that code, run `bin/console` for an interactive prompt.
-
-TODO: Delete this and the text above, and describe your gem
+This gem is a terminal based typing practice game.
 
 ## Installation
